Tweak abbreviations.
If the abbreviation is a prefix of the long form, don't put the
abbreviation in parenthesis. Eg.

\makeAbbreviation{\letter}{December Letter}{Letter sent 12/12/2012}
long form: Letter sent 12/12/2012 (December Letter)
short form: December Letter
comment: This is the normal abbreviation, the short form appears in
parenthesis after the long form.

\makeAbbreviation{\letter}{Letter}{Letter sent 12/12/2012}
long form: Letter sent 12/12/2012
short form: Letter
comment: This is the modification. When the short form is a prefix of
the long form, don't put the short form in parentheses.