Improve caseReference. Show Appendix references in parens.
Rename caserecord to caseReference. This name is more accurate because
what is being linked is a case's shortname with its reference.

Make Appendix an optional argument. This makes it easy to change the
name of the appendex later. (Although it needlessly complicates the
code now.)

Show all references to the appendix in parens. This unifies the
treatment of appendix references and makes it easier to format
automatically because parens are always closed with a closing paren
but comma-delimmited clauses can be closed with a period.