Fix hyphen in appendix references to page ranges.
When pinciting to a page range in a source that is in the appendix,
the hyphen in the page range in the reference to the appendix is
truncated to look more like a dot.

E.g., \pincite{Smith}{3-4} -> Smith at 3-4 (Appendix at 44-45). (The
hyphen between 44 and 45 looks like a dot.

Fix this by specifying an en dash.